requirements.txt
Django==4.2.7
environs==9.5.0
Pillow==10.1.0
phonenumberslite==8.13.25
phonenumbers==8.13.29
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
requests==2.31.0
